# docs/swagger/ecommerce_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "eCommerce API",
    "version": "1.0",
    "contact": {
      "name": "Phong ND",
      "url": "https://github.com/phongnd2802/go-ecommerce-microservices",
      "email": "duyphong02802@gmail.com"
    }
  },
  "tags": [
    {
      "name": "UserService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/users/register": {
      "post": {
        "summary": "Register",
        "description": "Use this API to register",
        "operationId": "UserService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbRegisterRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/v1/users/set_password": {
      "post": {
        "summary": "Update password register",
        "description": "Use this API to Update password register",
        "operationId": "UserService_UpdatePasswordRegister",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbSetPasswordResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbSetPasswordRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/v1/users/verify_otp": {
      "post": {
        "summary": "Verify OTP",
        "description": "Use this API to Verify OTP",
        "operationId": "UserService_VerifyOTP",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbVerifyOTPResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbVerifyOTPRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "definitions": {
    "pbRegisterRequest": {
      "type": "object",
      "properties": {
        "verifyKey": {
          "type": "string"
        }
      }
    },
    "pbRegisterResponse": {
      "type": "object",
      "properties": {
        "verifyId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbSetPasswordRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "pbSetPasswordResponse": {
      "type": "object",
      "properties": {
        "userEmail": {
          "type": "string"
        },
        "userNickname": {
          "type": "string"
        }
      }
    },
    "pbVerifyOTPRequest": {
      "type": "object",
      "properties": {
        "verifyKey": {
          "type": "string"
        },
        "verifyCode": {
          "type": "string"
        }
      }
    },
    "pbVerifyOTPResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}